+++ src/treeProvider.ts
@@ -1,15 +1,19 @@
 import * as vscode from 'vscode';
 import { Article, HackerNewsApi } from './hackerNewsApi';
-import path = require('path');
 import moment = require('moment');
 
 export class NodeDependenciesProvider implements vscode.TreeDataProvider<TreeItem> {
   private hackerNewsUrl = 'https://news.ycombinator.com/item?id=';
   private hackerNewsApi: HackerNewsApi = new HackerNewsApi();
+  private articleList: Article[] = [];
+  private history: History = new History();
 
-  constructor() { }
+  constructor(private context: vscode.ExtensionContext) { }
 
   async populateArticleTree(articles: Article[]): Promise<TreeItem[]> {
+
+    this.history.articlesRead = await this.context.globalState.get('articleHistory') || [];
+
     const tree: TreeItem[] = [];
     for (const article of articles) {
       const url = article.url ? article.url : `${this.hackerNewsUrl}${article.id}`;
@@ -25,18 +29,17 @@ export class NodeDependenciesProvider implements vscode.TreeDataProvider<TreeIte
       };
 
       const treeNode: TreeItem = new TreeItem(article.title, [childNode]);
+      const iconColor = this.isArticleRead(article.id) ? 'newsHack.read' : 'newsHack.unread';
       treeNode.tooltip = `${article.title} - ${url}`;
       treeNode.description = `${url}`;
-      treeNode.iconPath = new vscode.ThemeIcon('link');
+      treeNode.iconPath = new vscode.ThemeIcon('link', new vscode.ThemeColor(iconColor));
       treeNode.command = {
         command: 'hack-news.openArticle',
         title: 'Open Article',
-        arguments: [url],
+        arguments: [url, article.id],
       };
-
       tree.push(treeNode);
     }
-
     return tree;
   }
 
@@ -53,15 +56,39 @@ export class NodeDependenciesProvider implements vscode.TreeDataProvider<TreeIte
 
   async getChildren(element?: TreeItem): Promise<TreeItem[] | undefined> {
     if (element === undefined) {
-      return this.populateArticleTree(await this.hackerNewsApi.getTopStories());
+      this.articleList = await this.hackerNewsApi.getTopStories();
+      return this.populateArticleTree(this.articleList);
     }
 
     return Promise.resolve(element.children || undefined);
   }
 
-  // TODO: Dispose - need some cleanup for the extension.
+  isArticleRead(articleId: number): boolean {
+    if (this.history.articlesRead && this.history.articlesRead.length > 0) {
+      return this.history.articlesRead.includes(articleId);
+    }
+    return false;
+  }
+
+  async markArticleRead(articleId: number) {
+    this.history.articlesRead = await this.context.globalState.get('articleHistory') || [];
+    if (this.history.articlesRead) {
+      if (this.history.articlesRead.length > 200) {
+        this.history.articlesRead.shift();
+      }
+      this.history.articlesRead.push(articleId);
+      await this.context.globalState.update('articleHistory', this.history.articlesRead);
+    }
+  }
+
+  async clearArticleHistory() {
+    await this.context.globalState.update('articleHistory', []);
+  }
 }
 
+class History {
+  articlesRead: number[] | undefined = [];
+}
 class TreeItem extends vscode.TreeItem {
   children?: TreeItem[];