+++ CHANGELOG.md
@@ -1,17 +1,29 @@
 # Change Log
 
-All notable changes to the "hack-news" extension will be documented in this file.
-
+All notable changes to the "News-Hack" Vscode extension will be documented in this file.
 
 ### [Released]
 
+#### v0.0.8
+
+- Fixed missing clear article history button
+
+#### v0.0.7
+
+- Added article history (records up to 200 read articles)
+- Added button to clear article history
+- Misc fixes
+
 #### v0.0.5
+
 - Fixed duplicate refresh button
 - Added button to visit Hacker News website
 - Misc fixes to reduce install size of extension
 
 #### v0.0.4
+
 - Reformatted article comments to include points, author and comments
 
 #### v0.0.2
+
 - Initial release